requirements.txt
fastapi==0.115.14
httpx==0.28.1
pydantic-settings==2.10.1
uvicorn==0.35.0
cryptography==44.0.0

celery==5.5.0
celery[redis]

joblib==1.5.1
scikit-learn==1.7.0
catboost==1.2.8
xgboost==3.0.2
pandas==2.3.0
numpy==2.3.1
huggingface-hub==0.34.1

google-api-python-client==2.176.0
youtube-transcript-api==1.2.1
isodate==0.7.2